pyproject.toml
[project]
name = "spanishchatbot"
version = "0.1.0"
description = "A speech-to-speech pipeline that I can practice Spanish with"
authors = [
    {name = "ERIC KLINKHAMMER",email = "eklinkhammer@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "llvmlite (==0.44.0)",
    "numba (>=0.61.0,<0.62.0)",
    "openai-whisper (>=20240930,<20240931)",
    "numpy (>=1.24,<2.2)",
    "sounddevice (>=0.5.1,<0.6.0)",
    "scipy (>=1.15.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
whisper-transcribe = "src.spanishchatbot.main:main"
